requirements.txt
# ─── Web server ─────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ─── Core ML ────────────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
Pillow>=10.0.0
sentence-transformers>=2.6.0

# ─── OCR ────────────────────────────────────────────────────────────────────
easyocr>=1.7.1

# ─── Image processing ───────────────────────────────────────────────────────
opencv-python-headless>=4.9.0
numpy>=1.24.0

# ─── Utilities ──────────────────────────────────────────────────────────────
tqdm>=4.66.0
